refactor(mobile): split bill mutation
- extract payment and notification helpers
- add mutation source-contract test
